Reverting last PR breaking license table (#883)
In the last PR, I broke the license tables CTE's into their own individual tables. I also changed the license table to read these individual tables. The table has a 'unique' constraint on ID, and bad data trying to creep in is breaking the scheduled jobs.  We are still keeping the small tables, but reverting the license table to CTE's.

1) Impact of this is the dbt scheduled jobs will no longer fail. We will have to go back to see why we're getting bad data.
2) This fix is temporary until we figure out the source of the bad data.