Updating ARR transaction and rollfoward logic to take into account customer counts, SFDC migration issue and apply logic on fields not populated by Jim K (#911)
* Create arr_rollforward_test.sql

Hi Srikar this is a test

* Update arr_transactions.sql

Adding anniversary calculation beg_tenure_yr as an index for LTV calculation.  Revised label for former tenure_yr to be end_tenure_yr.  Clarifies that this is the index to market end year of customer

* Delete arr_rollforward_test.sql

* Update arr_transactions.sql

Updated for LTV tenure years adding tenure based on license start date and rename tenure year to be end tenure

* Updated for new logo field, addition of fiscal period conversion and references for original sources

* Creating ARR_TREND_TABLE

* Update schema.yml

* Update arr_transactions.sql

* Update arr_transactions.sql

* Bugfix added last_day function to fiscal quarter and fiscal year dates

* added last_day function on fiscal periods

* Major Update on Arr transaction and arr rollfoward logic taking into account initial migration issues with salesforce